requirements.txt
google-generativeai>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
category-encoders>=2.6.0
scipy>=1.11.0
